requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==1.0.0
attrs==24.3.0
babel==2.16.0
backports.tarfile==1.2.0
cassandra-driver==3.29.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docutils==0.21.2
frozenlist==1.5.0
geomet==0.2.1.post1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
joblib==1.4.2
kafka-python==2.0.2
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.15
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0
nh3==0.2.20
numpy==2.2.1
openmeteopy==0.1.0
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pkginfo==1.12.0
pluggy==1.5.0
propcache==0.2.1
pyarrow==19.0.0
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
streamlit==1.44.1
threadpoolctl==3.5.0
twine==6.0.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
zipp==3.21.0
